requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.4
garminconnect==0.2.19
garth==0.4.46
h11==0.14.0
idna==3.10
lxml==5.3.0
oauthlib==3.2.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
withings-sync==4.2.5
